docs: Use static html instead of php
The .html file extension is needed for offline docs, and configuring a webserver
to execute PHP in .html is as easy as changing the build system to support both.

But for a simple common header there's no need to reach for PHP. XSLT can do the
job. For ease of maintenance the header is read from an extra file.