Material objects dug free can be forced to be collected
The idea is that for some materials (e.g. earth chunks) it really
doesn't make sense to not immediately collect them - they would just
collide with something and recreate the material you just dug out.
So instead, this allows materials to specify that they want their
dug out objects to be collected immediately - if not, they simply
don't get created.

Note that this doesn't mean that material is lost, as it will simply
acumulate in the digger's material list instead. We might want to
cap that at some point though.