Savegames: Remove outdated assertion
Scenarios are now loaded with the same flags they were saved with. Thus,
the objectlist is told to skip player objects while loading. But this can
be safely ignored, because the scenario shouldn't contain any player
objects that could be ignored in the first place. But checking that loading
and saving is done with the same flag would be nice.